requirements.txt
pandas
torch==2.2.2
torchvision==0.17.2
sentence-transformers==2.7.0
transformers==4.40.1
Pillow==10.3.0
tqdm
numpy==1.26.4
langchain==0.1.20
langchain-core==0.1.52
langchain-google-genai==1.0.4
langchain-community==0.0.38
pinecone>=3.2.2
fastapi
uvicorn[standard]
python-dotenv
requests
gdown